chore: consolidate profile route
